www-apps/gitea: Add new package, version 1.20.2
The ebuild was based on the tree version earlier in 1.19.x series.
I dropped filecaps support, so it's necessary to configure direct
network access yourself (e.g. using systemd socket magic[1]) or use a
TLS termination proxy like nginx, caddy or traefik in front.

Besides, the gitea builds and runs fine now w/ PIE (or it is disabled
on a global scale somewhere). However, it's not necessary anymore to
strip out the flags in the ebuild[2].

[1] ListenStream in systemd socket configuration files
    and for a explanation e.g.
    https://liquidat.wordpress.com/2018/01/04/howto-run-programs-as-non-root-user-on-privileged-ports-via-systemd/
[2] https://bugs.gentoo.org/832756

Signed-off-by: Nils Freydank <nils.freydank@posteo.de>